fix: More test fixes
Updated the `FixedLengthFile` class in the `Archivist.DataTypes` namespace to include a line separator. Added two constructors, one default initializing the line separator to "\r\n", and another allowing the line separator to be specified. Also added a `LineSeparator` property to get the line separator used in the fixed-length file.

The `GetContent` method now uses the `LineSeparator` property instead of `Environment.NewLine` when joining the records.

Updated the `FixedLengthFile` instances in the `GetContent_ShouldReturnRecordsAsString` and `ToString_ShouldReturnRecordsAsString` test methods in the `FixedLengthFileTests` class to use the new constructor that accepts a line separator.

Improved readability of the `FixedLengthFile` class by updating comments for the comparison and equality operators and methods to span multiple lines.