fix: Add native Linux libraries to OCR project
The changes in the `Archivist.OCR.csproj` file include the addition of two new files to be included in the package. Specifically, the files `runtimes/linux-x64/native/libleptonica-1.82.0.so` and `runtimes/linux-x64/native/libtesseract50.so` are now being packed and included in the `runtimes/linux-x64/native/` directory of the package.